requirements.txt
anyio==3.6.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
black==22.3.0
boto3==1.24.17
botocore==1.27.17
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coverage==6.4.1
cryptography==37.0.2
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.15.0
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
jupyter-client==7.3.4
jupyter-core==4.10.0
mangum==0.15.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
moto==3.1.15
mypy-extensions==0.4.3
nest-asyncio==1.5.5
numpy==1.23.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pyzmq==23.2.0
requests==2.28.0
responses==0.21.0
s3transfer==0.6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
stack-data==0.3.0
starlette==0.19.1
tomli==2.0.1
tornado==6.1
traitlets==5.3.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.18.1
wcwidth==0.2.5
Werkzeug==2.1.2
xmltodict==0.13.0
